fix various TODOs in crate root
