+++ src/tests/mod.rs
@@ -19,6 +19,7 @@ mod background_tasks;
 mod display_name_verification;
 mod explicit;
 mod process_admin_cmds;
+mod live_mocker;
 
 // Convenience type
 pub type F = IdentityFieldValue;
